requirements.txt
pandas~=2.2.2
streamlit~=1.31.1
pillow~=10.3.0
matplotlib~=3.8.3
SQLAlchemy~=2.0.28
requests~=2.32.3
PyPDF2~=3.0.1
beautifulsoup4~=4.12.3